fix a dropped error in dispatch.iced
In commit f464c41 we factored out this line

    error = new Error(error) if typeof(error) is 'string'

into this function

    unwrap_incoming_error : (s) -> new Error(s) if typeof(s) is 'string'
    ...
    error = @unwrap_incoming_error error

What looked like a simple refactor was actually a tricky bug, because of
CoffeScript's interpretation of `if`. In the original code, `if` guarded
the *entire statement*, meaning that `error` was not overwritten if it
was not a string. However in the new code, the assignment became
unconditional, and `error` was set to `undefined` if it was not a
string. This fix is to have `unwrap_incoming_error` return the original
error in that case.